+++ CHANGELOG.md
@@ -11,6 +11,7 @@ The changelog format is based on [Keep a Changelog](https://keepachangelog.com/e
 ## Added
 
 - Configuration property `xatkit.react.public_url` to specify the path of the socket.io server. This property is used when accessing the bot using the `/admin` url, and allows to tune the HTML generation to connect to a specific react server. The provided URL must contain the hostname, port, and path to the react server, and can be used in nginx configuration for URL rewriting. If not specified the `/admin` HTML page will point to `http://localhost:<react port>`.
+- A set of `RestHandler` mimicking the behavior of the `/admin` endpoint previously defined in *xatkit-runtime* (see [#237](https://github.com/xatkit-bot-platform/xatkit-runtime/issues/237) for more information). The testing page can still be accessed at `localhost:5000/admin` (or the public URL specified in the configuration), and behave as before. **Note that these endpoints are only accessible when using the RestPlatform**.
 
 ### Changed